Patch for bug where exclude status doesn't save
Removed ep_set_option() function and replaced with  update_option()
from WP core. It is not necessary to delete, then re-add the option
because update_option() will update with new value if an option exists,
and create it if it does not.  This prevents the bug where if the
option does not get deleted for some reason, the new value will not get
saved.