Improvements to ep_update_exclusions()
Added code to prevent saving on auto-save, a check to be sure the
person updating the value has the rights to do so, and a check to
prevent saving the post revision IDs to the excluded pages array.
Previously it saved both the primary post ID and the ID of the
revision. This will cut the size of the excluded pages array by at
least 50%.