Removes any revision post IDs from DB when plugin is updated
I didn't want to verify the main post ID because with the way the old
plugin was coded revision IDs got stranded. We can not be positive that
the post is currently intended to be hidden just because there is a
revision ID in the array. If the post was excluded, then updated and
un-excluded, the revision ID would remain, but the post ID would be
gone. It is better to simply remove all revision IDs as they are not
what is being used for displaying the navigations/menus anyway.